Convert label tag handler to TS
