+++ examples/inspector/inspector.html
@@ -209,7 +209,7 @@ limitations under the License.
     <script src="../../build/ts/swf/parser/button.js"></script>
     <script src="../../build/ts/swf/parser/font.js"></script>
     <script src="../../src/swf/image.js"></script>
-    <script src="../../src/swf/label.js"></script>
+    <script src="../../build/ts/swf/parser/label.js"></script>
     <script src="../../src/swf/shape.js"></script>
     <script src="../../src/swf/sound.js"></script>
     <script src="../../src/swf/text.js"></script>
