+++ src/shumway.player.package
@@ -52,7 +52,7 @@ build/ts/swf/parser/bitmap.js
 build/ts/swf/parser/button.js
 build/ts/swf/parser/font.js
 src/swf/image.js
-src/swf/label.js
+build/ts/swf/parser/label.js
 src/swf/shape.js
 src/swf/sound.js
 src/swf/text.js
