+++ utils/builder/templates/scripts.js
@@ -43,7 +43,7 @@ load($SHUMWAY_ROOT + "build/ts/swf/parser/button.js");
 load($SHUMWAY_ROOT + "build/ts/swf/parser/font.js");
 
 load($SHUMWAY_ROOT + "src/swf/image.js");
-load($SHUMWAY_ROOT + "src/swf/label.js");
+load($SHUMWAY_ROOT + "build/ts/swf/parser/label.js");
 load($SHUMWAY_ROOT + "src/swf/shape.js");
 load($SHUMWAY_ROOT + "src/swf/sound.js");
 load($SHUMWAY_ROOT + "src/swf/text.js");
